build-fix: unbreak the full-game compile + link (InputBuffer namespace, gxMessageFilterFlags ODR)
Two pre-existing build breaks surfaced when wiring the Apt render TUs into the
full game build (both independent of the render work):

1. C2757 in every TU that bridges GUI + game-state (e.g. BrnMain): the GUI headers
   declared a global `class InputBuffer` (CgsGuiState.h as a scope for GuiEventQueue;
   CgsGuiShared.h for mpGDMInput) which clashed with the game-state's global
   `namespace InputBuffer` (BrnDriveThruManager.h + 8 siblings). The DWARF-correct
   model is a single global event-queue namespace InputBuffer, so the GUI side is now
   `namespace InputBuffer { class GuiEventQueue; }` and mpGDMInput is retyped to
   InputBuffer::GuiEventQueue* (the GameDataModule GUI-event input queue it really is).
   mpGDMInput/Set/GetGDMInput have no external callers, so the retype is contained.

2. LNK2019 on CgsDev::Message::gxMessageFilterFlags: CgsRemapDevice.cpp carried a stale
   local `extern u32` of it while the real definition (CgsLog.cpp) + canonical decls
   (CgsLog.h / CgsMessage.h FilterFlag) are u64 -- the u32 mangling never resolved.
   Corrected the local extern to u64.

With these, the full game build compiles 100% clean (BrnMain + all TUs) and links
except for the Apt render subsystem's own additive TUs (wired separately).

+++ src/GameShared/GameClasses/System/FileSystem/Devices/CgsRemapDevice.cpp
@@ -22,7 +22,7 @@
 
 namespace CgsDev
 {
-    namespace Message { extern u32 gxMessageFilterFlags; }   // defined in Development/Log/CgsLog.cpp
+    namespace Message { extern u64 gxMessageFilterFlags; }   // defined in Development/Log/CgsLog.cpp (u64 -- matches CgsLog.h / CgsMessage.h FilterFlag)
     namespace Log { void DebugPrint(const char* pcText, int liArg); void DebugPrint(const char*, int) { __debugbreak(); } }
 }
